Temperature Conversion Program
A program written in C++ that allows the user to choose whether they would like to enter a temperature in Fahrenheit or Celsius or quit the program. The program converts the temperature entered, allows the user to quit, or prints an error message stating the wrong input has been entered.